Add files via upload
Initial commit.  Includes the ability to browse to the old zip and a new zip.  It will compare the lists of mods in both versions to determine which mods have been added, removed or updated.  It will also report changes to the mod loader.

It doesn't do anything special with "overrides" at this time